fixing menu hanups

+++ go/desktop_win_bitmapimage.go
@@ -81,9 +81,9 @@ func (m *BitmapImage) Height() int {
 	return m.image.Bounds().Max.Y - m.image.Bounds().Min.Y
 }
 
-func (m *BitmapImage) Draw(x LONG, y LONG,hdcDst HDC) {
-	cx := m.Width()
-	cy := m.Height()
+func (m *BitmapImage) Draw(x LONG, y LONG, hdcDst HDC) {
+	cx := LONG(m.Width())
+	cy := LONG(m.Height())
 	
     hdcSrc := HDCPtr(CreateCompatibleDC.Call(Arg(hdcDst)))
 	if hdcSrc == 0 {
